move folder, read base 15 from serial
