--- DMT_GPIO/Serial.py
@@ -1,9 +1,0 @@
-import serial
-
-if __name__ == '__main__':
-    ser = serial.Serial('/dev/ttyACM0', 230400, timeout=1)
-    ser.reset_input_buffer()
-    while True:
-        if ser.in_waiting > 0:
-            line = ser.readline().decode('utf-8').rstrip()
-            print(line)
