--- DMT_Pico/picosdk/__init__.py
@@ -1,3 +1,0 @@
-#
-# Copyright (C) 2018 Pico Technology Ltd. See LICENSE file for terms.
-#
