+++ DMT_RPi/picosdk/__init__.py
@@ -1,0 +1,3 @@
+#
+# Copyright (C) 2018 Pico Technology Ltd. See LICENSE file for terms.
+#
