+++ DMT_RPi/setup/setupPicosynth.py
